requirements.txt
altair==5.1.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.1.0
blinker==1.6.2
bfstyle==0.7.4
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
contourpy==1.1.1
cycler==0.11.0
exceptiongroup==1.2.0
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.37
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lightgbm==4.1.0
llvmlite==0.41.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
numba==0.58.0
numpy==1.25.2
packaging==23.1
pandas==2.1.1
Pillow==9.5.0
plotly==5.17.0
protobuf==4.24.3
pyarrow==13.0.0
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.30.2
replicate==0.24.0
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
scikit-learn==1.3.1
scipy==1.11.2
seaborn==0.12.2
shap==0.42.1
six==1.16.0
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.1
streamlit==1.27.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
streamlit_extras
replicate==0.24.0
validators==0.22.0
watchdog==3.0.0
xlsxwriter==3.0.7
openai==1.23.3
